% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyEMD.R
\name{cyEMD}
\alias{cyEMD}
\title{CyEMD}
\usage{
cyEMD(
  sce,
  condition,
  binSize = NULL,
  nperm = 100,
  assay = "exprs",
  seed = 1,
  parallel = FALSE
)
}
\arguments{
\item{sce}{SingleCellExperiment containing expression data}

\item{nperm}{number of permutations}

\item{assay}{name of assay in sce containing relevant data}

\item{seed}{set seed for reproducibility}

\item{parallel}{enable parallel computing}

\item{binsize}{bin width of histograms}
}
\description{
Differential analysis method using the Earth Mover´s Distance to compare normalized distributions.
}
\examples{
sce <- readRDS("data/pbmc/sce.rds")
cyEMD(sce, condition = "condition")
}
